other-agent-files/file-agent/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
asyncpg
openai